move customer repository to Orckestra.Composer
