+++ src/Orckestra.Composer.MyAccount/MyAccountPlugin.cs
@@ -1,5 +1,6 @@
 ﻿using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
+using Orckestra.Composer.Repositories;
 
 namespace Orckestra.Composer.MyAccount
 {
@@ -8,12 +9,9 @@ namespace Orckestra.Composer.MyAccount
         public void Register(IComposerHost host)
         {
             host.Register<MembershipViewService, IMembershipViewService>();
-
             host.Register<CustomerRepository, ICustomerRepository>();
             host.Register<CustomerAddressRepository, ICustomerAddressRepository>();
-
             host.Register<CustomerAddressViewService, ICustomerAddressViewService>();
-
             host.Register<CustomerViewService, ICustomerViewService>();
 
             host.MetadataRegistry.LoadViewModelMetadataInAssemblyOf(typeof(MyAccountPlugin).Assembly);
