+++ src/Orckestra.Composer.MyAccount/Services/CustomerViewService.cs
@@ -5,12 +5,13 @@ using System.Linq;
 using System.Threading.Tasks;
 using Orckestra.Composer.Enums;
 using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.ViewModels;
+using Orckestra.Composer.Repositories;
 using Orckestra.Composer.Providers;
 using Orckestra.Composer.Providers.Localization;
 using Orckestra.Composer.Services;
 using Orckestra.Composer.ViewModels;
+using Orckestra.Composer.Parameters;
 using Orckestra.Overture.ServiceModel.Customers;
 using static Orckestra.Composer.Utils.MessagesHelper.ArgumentException;
