+++ src/Orckestra.Composer.MyAccount/Services/MembershipViewService.cs
@@ -5,8 +5,8 @@ using System.Threading.Tasks;
 using System.Web.Security;
 using Orckestra.Composer.MyAccount.Parameters;
 using Orckestra.Composer.MyAccount.Providers;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.ViewModels;
+using Orckestra.Composer.Repositories;
 using Orckestra.Composer.Parameters;
 using Orckestra.Composer.Providers;
 using Orckestra.Composer.Providers.Membership;
