+++ src/Orckestra.Composer/Repositories/CustomerRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Orckestra.Composer.Configuration;
 using Orckestra.Composer.Exceptions;
-using Orckestra.Composer.MyAccount.Parameters;
+using Orckestra.Composer.Parameters;
 using Orckestra.Overture;
 using Orckestra.Overture.Caching;
 using Orckestra.Overture.ServiceModel.Customers;
@@ -13,7 +13,7 @@ using Orckestra.Overture.ServiceModel.Requests.Customers.Membership;
 using static Orckestra.Composer.Utils.MessagesHelper.ArgumentException;
 
 
-namespace Orckestra.Composer.MyAccount.Repositories
+namespace Orckestra.Composer.Repositories
 {
     /// <summary>
     /// Repository for interacting with Customers
