+++ tests/Orckestra.Composer.MyAccount.Tests/Orckestra.Composer.MyAccount.Tests.csproj
@@ -165,7 +165,6 @@
     <Compile Include="Mock\MockAddressFactory.cs" />
     <Compile Include="Mock\MockMembershipProviderFactory.cs" />
     <Compile Include="Mock\MockMembershipFactory.cs" />
-    <Compile Include="Mock\MockCustomerFactory.cs" />
     <Compile Include="Mock\MockHttpContext.cs" />
     <Compile Include="Mock\MockComposerContextFactory.cs" />
     <Compile Include="Mock\MockMyAccountUrlProviderFactory.cs" />
@@ -173,14 +172,7 @@
     <Compile Include="Repositories\CustomerAddressRepository_CreateAddressAsync.cs" />
     <Compile Include="Repositories\CustomerAddressRepository_DeleteAddressAsync.cs" />
     <Compile Include="Repositories\CustomerAddressRepository_UpdateAddressAsync.cs" />
-    <Compile Include="Repositories\CustomerRepository_ChangePasswordAsync.cs" />
-    <Compile Include="Repositories\CustomerRepository_ResetPasswordAsync.cs" />
-    <Compile Include="Repositories\CustomerRepository_SendResetPasswordInstructionsAsync.cs" />
-    <Compile Include="Repositories\CustomerRepository_CreateUserAsync.cs" />
     <Compile Include="Repositories\CustomerAddressRepository_GetCustomerAddressesAsync.cs" />
-    <Compile Include="Repositories\CustomerRepository_GetCustomerByIdAsync.cs" />
-    <Compile Include="Repositories\CustomerRepository_GetCustomerByTicketAsync.cs" />
-    <Compile Include="Repositories\CustomerRepository_UpdateUserAsync.cs" />
     <Compile Include="Services\CustomerAddressViewService_CreateAddressAsync.cs" />
     <Compile Include="Services\CustomerAddressViewService_DeleteAddressAsync.cs" />
     <Compile Include="Services\CustomerAddressViewService_GetAddressListViewModel.cs" />
@@ -216,6 +208,10 @@
     <ProjectReference Include="..\..\src\Orckestra.Composer\Orckestra.Composer.csproj">
       <Project>{D00BD58B-93FE-4F3D-BBC8-6F7873CA6830}</Project>
       <Name>Orckestra.Composer</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Orckestra.Composer.Tests\Orckestra.Composer.Tests.csproj">
+      <Project>{24eaec4f-3e3c-4a4a-ab18-f8d64a7c0673}</Project>
+      <Name>Orckestra.Composer.Tests</Name>
     </ProjectReference>
     <ProjectReference Include="..\Orckestra.ForTests\Orckestra.ForTests.csproj">
       <Project>{B2BE051E-82D7-4643-BB56-98F161833E30}</Project>
