+++ tests/Orckestra.Composer.MyAccount.Tests/Services/CustomerAddressViewService_GetCreateAddressViewModelAsync.cs
@@ -6,11 +6,13 @@ using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Parameters;
 using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
 using Orckestra.Composer.MyAccount.Tests.Mock;
 using Orckestra.Composer.MyAccount.ViewModels;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.ForTests;
 using Orckestra.ForTests.Mock;
