+++ tests/Orckestra.Composer.MyAccount.Tests/Services/CustomerAddressViewService_GetEditAddressViewModel.cs
@@ -8,12 +8,13 @@ using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
 using Orckestra.Composer.Country;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Parameters;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
 using Orckestra.Composer.MyAccount.Tests.Mock;
 using Orckestra.Composer.MyAccount.ViewModels;
-using Orckestra.Composer.Repositories;
 using Orckestra.ForTests;
 using Orckestra.ForTests.Mock;
