+++ tests/Orckestra.Composer.MyAccount.Tests/Services/CustomerAddressViewService_SetDefaultAddressAsync.cs
@@ -12,6 +12,7 @@ using Orckestra.Composer.MyAccount.Services;
 using Orckestra.Composer.MyAccount.Tests.Mock;
 using Orckestra.Composer.MyAccount.ViewModels;
 using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Parameters;
 using Orckestra.ForTests.Mock;
 using Orckestra.Overture.ServiceModel;
 using Orckestra.Overture.ServiceModel.Customers;
