+++ tests/Orckestra.Composer.MyAccount.Tests/Services/CustomerViewService_GetAccountHeaderViewModelAsync.cs
@@ -5,11 +5,13 @@ using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Parameters;
 using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
 using Orckestra.Composer.MyAccount.Tests.Mock;
 using Orckestra.Composer.MyAccount.ViewModels;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.ForTests;
 using Orckestra.ForTests.Mock;
 using System.Threading.Tasks;
