+++ tests/Orckestra.Composer.MyAccount.Tests/Services/CustomerViewService_UpdateAccountAsync.cs
@@ -6,8 +6,10 @@ using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Parameters;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
 using Orckestra.Composer.MyAccount.Tests.Mock;
 using Orckestra.Composer.MyAccount.ViewModels;
