+++ tests/Orckestra.Composer.MyAccount.Tests/Services/MembershipViewService_ForgotPasswordAsync.cs
@@ -7,8 +7,9 @@ using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
 using Orckestra.Composer.Exceptions;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Parameters;
 using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
 using Orckestra.ForTests;
