+++ tests/Orckestra.Composer.MyAccount.Tests/Services/MembershipViewService_GetChangePasswordViewModelAsync.cs
@@ -8,10 +8,12 @@ using Moq.AutoMock;
 using NUnit.Framework;
 using Orckestra.Composer.MyAccount.Parameters;
 using Orckestra.Composer.MyAccount.Providers;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
 using Orckestra.Composer.MyAccount.Tests.Mock;
 using Orckestra.Composer.MyAccount.ViewModels;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Parameters;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.ForTests;
 using Orckestra.ForTests.Mock;
 using System.Threading.Tasks;
