+++ tests/Orckestra.Composer.MyAccount.Tests/Services/MembershipViewService_LoginAsync.cs
@@ -9,9 +9,10 @@ using Moq.AutoMock;
 using NUnit.Framework;
 using Orckestra.Composer.MyAccount.Parameters;
 using Orckestra.Composer.MyAccount.Providers;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
-using Orckestra.Composer.MyAccount.Tests.Mock;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Parameters;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.ForTests;
 using Orckestra.Overture.ServiceModel.Customers;
 using static Orckestra.Composer.Utils.MessagesHelper.ArgumentException;
