+++ tests/Orckestra.Composer.MyAccount.Tests/Services/MembershipViewService_RegisterAsync.cs
@@ -11,12 +11,13 @@ using NUnit.Framework;
 using Orckestra.Composer.Exceptions;
 using Orckestra.Composer.MyAccount.Parameters;
 using Orckestra.Composer.MyAccount.Providers;
-using Orckestra.Composer.MyAccount.Repositories;
 using Orckestra.Composer.MyAccount.Services;
 using Orckestra.Composer.MyAccount.Tests.Mock;
 using Orckestra.Composer.MyAccount.ViewModels;
 using Orckestra.Composer.Parameters;
 using Orckestra.Composer.Providers;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.ForTests;
 using Orckestra.ForTests.Mock;
 using Orckestra.Overture.ServiceModel.Customers;
