+++ tests/Orckestra.Composer.MyAccount.Tests/Services/MembershipViewService_ResetPasswordAsync.cs
@@ -9,9 +9,10 @@ using Moq.AutoMock;
 using NUnit.Framework;
 using Orckestra.Composer.MyAccount.Parameters;
 using Orckestra.Composer.MyAccount.Providers;
-using Orckestra.Composer.MyAccount.Repositories;
+using Orckestra.Composer.Repositories;
 using Orckestra.Composer.MyAccount.Services;
 using Orckestra.Composer.MyAccount.Tests.Mock;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.ForTests;
 
 namespace Orckestra.Composer.MyAccount.Tests.Services
