+++ tests/Orckestra.Composer.Tests/Repositories/CustomerRepository_CreateUserAsync.cs
@@ -7,16 +7,16 @@ using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
-using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
-using Orckestra.Composer.MyAccount.Tests.Mock;
+using Orckestra.Composer.Parameters;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.ForTests;
 using Orckestra.Overture;
 using Orckestra.Overture.ServiceModel.Customers;
 using Orckestra.Overture.ServiceModel.Requests.Customers.Membership;
 using static Orckestra.Composer.Utils.ExpressionUtility;
 
-namespace Orckestra.Composer.MyAccount.Tests.Repositories
+namespace Orckestra.Composer.Tests.Repositories
 {
     // ReSharper disable once InconsistentNaming
     class CustomerRepository_CreateUserAsync
