+++ tests/Orckestra.Composer.Tests/Repositories/CustomerRepository_GetCustomerByIdAsync.cs
@@ -6,17 +6,16 @@ using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
-using Orckestra.Composer.Configuration;
-using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
-using Orckestra.Composer.MyAccount.Tests.Mock;
+using Orckestra.Composer.Parameters;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.ForTests;
 using Orckestra.Overture;
 using Orckestra.Overture.Caching;
 using Orckestra.Overture.ServiceModel.Customers;
 using Orckestra.Overture.ServiceModel.Requests.Customers;
 
-namespace Orckestra.Composer.MyAccount.Tests.Repositories
+namespace Orckestra.Composer.Tests.Repositories
 {
     // ReSharper disable once InconsistentNaming
     internal class CustomerRepository_GetCustomerByIdAsync
