+++ tests/Orckestra.Composer.Tests/Repositories/CustomerRepository_GetCustomerByTicketAsync.cs
@@ -4,12 +4,12 @@ using FluentAssertions;
 using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
-using Orckestra.Composer.MyAccount.Repositories;
-using Orckestra.Composer.MyAccount.Tests.Mock;
+using Orckestra.Composer.Repositories;
+using Orckestra.Composer.Tests.Mock;
 using Orckestra.Overture;
 using Orckestra.Overture.ServiceModel.Requests.Customers;
 
-namespace Orckestra.Composer.MyAccount.Tests.Repositories
+namespace Orckestra.Composer.Tests.Repositories
 {
     // ReSharper disable once InconsistentNaming
     internal class CustomerRepository_GetCustomerByTicketAsync
