+++ tests/Orckestra.Composer.Tests/Repositories/CustomerRepository_SendResetPasswordInstructionsAsync.cs
@@ -5,13 +5,13 @@ using Moq;
 using Moq.AutoMock;
 using NUnit.Framework;
 using Orckestra.Composer.Exceptions;
-using Orckestra.Composer.MyAccount.Parameters;
-using Orckestra.Composer.MyAccount.Repositories;
+using Orckestra.Composer.Parameters;
+using Orckestra.Composer.Repositories;
 using Orckestra.Overture;
 using Orckestra.Overture.ServiceModel.Customers.Membership;
 using Orckestra.Overture.ServiceModel.Requests.Customers.Membership;
 
-namespace Orckestra.Composer.MyAccount.Tests.Repositories
+namespace Orckestra.Composer.Tests.Repositories
 {
     // ReSharper disable once InconsistentNaming
     internal class CustomerRepository_SendResetPasswordInstructionsAsync
